+++ src/cosmosdb-preview/azext_cosmosdb_preview/custom.py
@@ -40,6 +40,12 @@ from azext_cosmosdb_preview.vendored_sdks.azure_mgmt_cosmosdb.models import (
     MongoDBDatabaseCreateUpdateParameters,
     MongoDBCollectionResource,
     MongoDBCollectionCreateUpdateParameters,
+    TableResource,
+    TableCreateUpdateParameters,
+    GremlinDatabaseResource,
+    GremlinDatabaseCreateUpdateParameters,
+    GremlinGraphResource,
+    GremlinGraphCreateUpdateParameters,
     Location,
     CreateMode,
     ConsistencyPolicy,
@@ -1792,3 +1798,142 @@ def cli_begin_redistribute_mongo_container_partition_throughput(client,
                                                                                                              redistribute_throughput_parameters=redistribute_throughput_parameters)
 
     return async_partition_redistribute_throughput_result.result()
+
+def cli_cosmosdb_gremlin_database_restore(cmd,
+                                          client,
+                                          resource_group_name,
+                                          account_name,
+                                          database_name,
+                                          restore_timestamp=None):
+    restorable_database_accounts_client = cf_restorable_database_accounts(cmd.cli_ctx, [])
+    restorable_database_accounts = restorable_database_accounts_client.list()
+    restorable_database_accounts_list = list(restorable_database_accounts)
+    restore_timestamp_datetime_utc = _convert_to_utc_timestamp(restore_timestamp)
+    restorable_database_account = None
+
+    for account in restorable_database_accounts_list:
+        if account.account_name == account_name:
+            if account.deletion_time is not None:
+                if account.deletion_time >= restore_timestamp_datetime_utc >= account.creation_time:
+                    raise CLIError("Cannot perform inaccount restore on a deleted database account {}".format(account_name))
+            else:
+                if restore_timestamp_datetime_utc >= account.creation_time:
+                    restorable_database_account = account
+                    break
+
+    if restorable_database_account is None:
+        raise CLIError("Cannot find a Gremlin database account with name {} that is online at {}".format(account_name, restore_timestamp))
+
+    # """Restores the deleted Azure Cosmos DB Gremlin database"""
+    create_mode = CreateMode.restore.value
+    restore_parameters = RestoreParameters(
+        restore_source=restorable_database_account.id,
+        restore_timestamp_in_utc=restore_timestamp
+    )
+
+    gremlin_database_resource = GremlinDatabaseCreateUpdateParameters(
+        resource=SqlDatabaseResource(
+            id=database_name,
+            create_mode=create_mode,
+            restore_parameters=restore_parameters)
+    )
+
+    return client.begin_create_update_gremlin_database(resource_group_name,
+                                                       account_name,
+                                                       database_name,
+                                                       gremlin_database_resource)
+
+
+def cli_cosmosdb_gremlin_graph_restore(cmd,
+                                       client,
+                                       resource_group_name,
+                                       account_name,
+                                       database_name,
+                                       graph_name,
+                                       restore_timestamp=None):
+    # """Restores the deleted Azure Cosmos DB Gremlin graph """
+    restorable_database_accounts_client = cf_restorable_database_accounts(cmd.cli_ctx, [])
+    restorable_database_accounts = restorable_database_accounts_client.list()
+    restorable_database_accounts_list = list(restorable_database_accounts)
+    restore_timestamp_datetime_utc = _convert_to_utc_timestamp(restore_timestamp)
+    restorable_database_account = None
+
+    for account in restorable_database_accounts_list:
+        if account.account_name == account_name:
+            if account.deletion_time is not None:
+                if account.deletion_time >= restore_timestamp_datetime_utc >= account.creation_time:
+                    raise CLIError("Cannot perform inaccount restore on a deleted gremlin database account {}".format(account_name))
+            else:
+                if restore_timestamp_datetime_utc >= account.creation_time:
+                    restorable_database_account = account
+                    break
+
+    if restorable_database_account is None:
+        raise CLIError("Cannot find a database account with name {} that is online at {}".format(account_name, restore_timestamp))
+
+    # """Restores the deleted Azure Cosmos DB Gremlin graph"""
+    create_mode = CreateMode.restore.value
+    restore_parameters = RestoreParameters(
+        restore_source=restorable_database_account.id,
+        restore_timestamp_in_utc=restore_timestamp
+    )
+
+    gremlin_graph_resource = GremlinGraphResource(
+        id=graph_name,
+        create_mode=create_mode,
+        restore_parameters=restore_parameters)
+
+    gremlin_graph_create_update_resource = GremlinGraphCreateUpdateParameters(
+        resource=gremlin_graph_resource,
+        options={})
+
+    return client.begin_create_update_gremlin_graph(resource_group_name,
+                                                    account_name,
+                                                    database_name,
+                                                    graph_name,
+                                                    gremlin_graph_create_update_resource)
+
+
+def cli_cosmosdb_table_restore(cmd,
+                               client,
+                               resource_group_name,
+                               account_name,
+                               table_name,
+                               restore_timestamp=None):
+    # """Restores the deleted Azure Cosmos DB Table"""
+    restorable_database_accounts_client = cf_restorable_database_accounts(cmd.cli_ctx, [])
+    restorable_database_accounts = restorable_database_accounts_client.list()
+    restorable_database_accounts_list = list(restorable_database_accounts)
+    restore_timestamp_datetime_utc = _convert_to_utc_timestamp(restore_timestamp)
+    restorable_database_account = None
+
+    for account in restorable_database_accounts_list:
+        if account.account_name == account_name:
+            if account.deletion_time is not None:
+                if account.deletion_time >= restore_timestamp_datetime_utc >= account.creation_time:
+                    raise CLIError("Cannot perform inaccount restore on a deleted table {}".format(account_name))
+            else:
+                if restore_timestamp_datetime_utc >= account.creation_time:
+                    restorable_database_account = account
+                    break
+
+    if restorable_database_account is None:
+        raise CLIError("Cannot find a account with name {} that is online at {}".format(account_name, restore_timestamp))
+
+    # """Restores the deleted Azure Cosmos DB Table"""
+    create_mode = CreateMode.restore.value
+    restore_parameters = RestoreParameters(
+        restore_source=restorable_database_account.id,
+        restore_timestamp_in_utc=restore_timestamp
+    )
+
+    table_resource = TableCreateUpdateParameters(
+        resource=TableResource(id=table_name,
+                               create_mode=create_mode,
+                               restore_parameters=restore_parameters),
+        options={})
+
+    return client.begin_create_update_table(resource_group_name,
+                                            account_name,
+                                            table_name,
+                                            table_resource)